moving 2.0 tests to a different file
